profiling: Support for forwarding PMU events as faults
Adds a flag in the TCB to denote that a process is being "profiled".
When this flag is set, on reception of a PMU IRQ, the kernel will send
a fault to the fault handler of that thread. If this flag is not set,
and a PMU IRQ is received, the IRQ is delivered as a notification to
the appropriate IRQ cap.

Signed-off-by: Krishnan Winter <krishnan-git@wintermail.me>