aarch64: use fixed virtual addr for kernel image
This matches RISC-V which also does this.

This is a precursor to some other boot protocol changes as a part of
the multikernel work where it is useful to be able to load the kernel
at different physical addresses using the same kernel image.

Previously, the kernel would run in the kernel physical memory window
at a device-specific virtual address.

Signed-off-by: julia <git.ts@trainwit.ch>