aarch64: kernel ELF memory aliasing requirements
As a result of the fixed virtual address space layout for the
kernel ELF, we have memory aliasing of the physical addresses
between the ELF and physical memory window. Document the
architectural requirements as we know it in case of future
modifications.

I have checked the users of paddr_to_pptr or ptrFromPaddr and
believe that all of them are used on userspace memory only.

Signed-off-by: julia <git.ts@trainwit.ch>